+++ database/repositories/tyre_impression_processing_repository.py
@@ -8,3 +8,7 @@ class TyreImpressionProcessingRepository(BaseRepository[TyreImpressionProcessing
         super().__init__(TyreImpressionProcessing)
 
 
+    def get_by_tyre_impression_id(self, tyre_impression_id: int) -> TyreImpressionProcessing:
+        return self.db.query(self.model).filter(self.model.tyre_impression_id == tyre_impression_id).first()
+
+