+++ preprocessing/__init__.py
@@ -1,0 +1,1 @@
+from .tyre_impression_processing_pipeline import TyreImpressionProcessingPipeline